requirements-cmsc720.txt
huggingface_hub[hf_transfer] @ git+https://github.com/huggingface/huggingface_hub
pytorch-fast-transformers
wandb
